requirements.txt
apache-airflow-providers-celery
apache-airflow-providers-sftp